--- a/dashboard/app.py
+++ b/dashboard/app.py
@@ -9,7 +9,6 @@ import plotly.express as px
 import plotly.graph_objects as go
 from datetime import datetime, timedelta
 import time
-from pathlib import Path
 
 # Page configuration
 st.set_page_config(