--- a/dashboard/services/performance_config.py
+++ b/dashboard/services/performance_config.py
@@ -6,8 +6,7 @@ import streamlit as st
 from functools import wraps
 import time
 import hashlib
-import json
-from typing import Any, Callable, Optional
+from typing import Any, Callable
 from datetime import datetime, timedelta