--- a/dashboard/test_connection.py
+++ b/dashboard/test_connection.py
@@ -2,7 +2,6 @@
 """Test dashboard connections"""
 
 import requests
-import json
 
 print("Testing Dashboard Connections...")
 print("=" * 50)