--- a/debug_ci_deployment.py
+++ b/debug_ci_deployment.py
@@ -8,7 +8,6 @@ import subprocess
 import json
 
 # import yaml  # Optional - will work without it
-import time
 from typing import Dict, Any