--- a/scripts/requirements.txt
+++ b/scripts/requirements.txt
@@ -1,0 +1,2 @@
+aiohttp>=3.8.0
+requests>=2.28.0