… .github/workflows/archived/auto-fix-ci.yml
… .github/workflows/archived/ci-auto-fix.yml
… .github/workflows/archived/ci-autofix-trigger.yml
… .github/workflows/archived/ci-hybrid.yml
… .github/workflows/archived/ci-monitor-trigger.yml
… .github/workflows/archived/docker-ci-optimized.yml
… .github/workflows/archived/performance-monitoring.yml
… .github/workflows/dev-ci-simplified.yaml
… .github/workflows/dev-ci-v2.yml
… .github/workflows/dev-ci.yml
… .github/workflows/docker-ci.yml
… .github/workflows/pr-merge-analysis.yml
… .github/workflows/pr-value-analysis.yml
… .github/workflows/quick-ci.yml
… AI_MONITORING_IMPLEMENTATION.md
… CI_OPTIMIZATION_ANALYSIS.md
… FINAL_MERGE_STATUS.md
… INTEGRATION_GUIDE.md
… TEST_COVERAGE_SUMMARY.md
… WEEK2_MERGE_CHECKLIST.md
… chart/templates/celery-worker.yaml
… chart/templates/fake-threads.yaml
… chart/templates/migrations-job.yaml
… chart/templates/postgres-simple.yaml
… chart/templates/postgres.yaml
… chart/templates/viral-metrics.yaml
… chart/values-ci-fast.yaml
… chart/values-ci-no-probes.yaml
… chart/values-ci.yaml
… chart/values-dev-fast.yaml